feat: add custom ids to shrts
