+++ types/index.ts
@@ -64,7 +64,6 @@ export interface ShrtDocument {
   isArchived: boolean;
   url: string;
   shrt_url: string | null;
-  title?: string;
   clicks?: number;
 }
